Improvements to follow best practices.
Creates a separate Root, Type, and BaseFormat file now.
Root has no namespace.

Added comments to the files to provide more guidance to schema authors.

Change default to flat style (which was already stated in README.md)

Improve README.md with links to DFDL schema style guide, which is a wiki
page for now, but someday will be on daffodil-site.

Improve README.md with link to daffodil-sbt plugin for more info about
the capabilities that offers for compiling, packaging.

The package string is added to prefix of namespace URI.

DAFFODIL-2976